The big folder restructuring of 2020
