--- Pipfile
@@ -1,17 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-
-[packages]
-flask = "*"
-flask-wtf = "*"
-flask-bootstrap = "*"
-flask-assets = "*"
-cssmin = "*"
-jinja = "*"
-
-[requires]
-python_version = "3.8"
