+++ api/pytest.ini
@@ -1,0 +1,6 @@
+[pytest]
+log_cli = 1
+log_cli_level = INFO
+markers =
+  cluster: marks test that require a k8s cluster in their environment
+  container: marks test that require the container environment
