+++ api/requirements.txt
@@ -1,0 +1,18 @@
+-i https://pypi.org/simple
+click==7.1.1
+cssmin==0.2.0
+dominate==2.5.1
+flask-assets==2.0
+flask-bootstrap==3.3.7.1
+flask-wtf==0.14.3
+Flask-RESTful==0.3.8
+flask==1.1.2
+pytest==6.0.1
+itsdangerous==1.1.0
+json_log_formatter==0.3.0
+jinja2==2.11.1
+markupsafe==1.1.1
+visitor==0.1.3
+webassets==2.0
+werkzeug==1.0.1
+wtforms==2.2.1
