+++ api/setup.cfg
@@ -1,0 +1,4 @@
+[coverage:run]
+branch = True
+source =
+    app
