+++ chart/kubeseal-webgui/Chart.yaml
@@ -1,5 +1,5 @@
 apiVersion: v1
 name: kubeseal-webgui
 description: A Helm chart for installing kubeseal-webgui
-version: 1.0.2
-appVersion: 1.0.2
+version: 2.0.0
+appVersion: 2.0.0
