--- conftest.py
@@ -1,18 +1,0 @@
-import pytest
-from app import create_app
-
-@pytest.fixture
-def app():
-    app = create_app({
-        'TESTING': True,
-    })
-
-    yield app
-
-@pytest.fixture
-def client(app):
-    return app.test_client()
-
-@pytest.fixture
-def runner(app):
-    return app.test_cli_runner()
