--- pytest.ini
@@ -1,4 +1,0 @@
-[pytest]
-markers =
-  cluster: marks test that require a k8s cluster in their environment
-  container: marks test that require the container environment
